requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.35.0
torch==2.1.0
pydantic==2.4.2
numpy==1.24.3
spacy==3.7.2
python-multipart==0.0.6
python-jose==3.3.0
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3 